requirements.txt
backports.lzma==0.0.14
dataclasses
joblib==1.0.1
numpy==1.19.5
Pillow==8.2.0
protobuf==3.15.8
PyYAML==5.4.1
scikit-learn==0.24.1
scipy==1.5.4
six==1.15.0
tensorboardX==2.2
threadpoolctl==2.1.0
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.7.4.3
pylint==2.8.2
